[FE] rename: 파일이름 및 폴더 구조 변경
- GroupCreate -> CreateGroup
- GroupCreatePage 컴포넌트 index.tsx로 파일이름 변경
- InputComponent(카테고리, 인원수, 그룹이름...) 폴더로 묶고 index.tsx를 export하도록 구조변경
- issue (#22)
