+++ client/src/pages/CreateGroupPage/GageBar.tsx
