+++ client/src/pages/CreateGroupPage/GroupInfo/index.tsx
